upgrade to use uv instead of conda

+++ docker/morphocluster/run.sh
@@ -1,10 +1,16 @@
-#!/bin/sh
+#!/bin/bash
 
-cp /authorized_keys /root/.ssh/authorized_keys
-chmod 600 /root/.ssh/authorized_keys
+# Handle authorized_keys - create empty file if directory doesn't exist
+if [ -f /authorized_keys ]; then
+    cp /authorized_keys /root/.ssh/authorized_keys
+    chmod 600 /root/.ssh/authorized_keys
+else
+    touch /root/.ssh/authorized_keys
+    chmod 600 /root/.ssh/authorized_keys
+fi
 
-. /opt/conda/etc/profile.d/conda.sh
-conda activate morphocluster
+# Activate Python virtual environment
+source /opt/venv/bin/activate
 
 export FLASK_APP=morphocluster
 export MORPHOCLUSTER_SETTINGS=config_docker.py
